+++ ros1/src/hardware/interface/march_hardware/test/mocks/mock_imotioncube.h
@@ -26,17 +26,17 @@ public:
 
   MOCK_METHOD0(isIncrementalEncoderMorePrecise, bool());
 
-  MOCK_METHOD0(getIncrementalPosition, double());
-  MOCK_METHOD0(getAbsolutePosition, double());
-  MOCK_METHOD0(getIncrementalVelocity, double());
-  MOCK_METHOD0(getAbsoluteVelocity, double());
+  MOCK_METHOD0(getIncrementalPosition, float());
+  MOCK_METHOD0(getAbsolutePosition, float());
+  MOCK_METHOD0(getIncrementalVelocity, float());
+  MOCK_METHOD0(getAbsoluteVelocity, float());
 
   MOCK_METHOD0(getMotorControllerVoltage, float());
   MOCK_METHOD0(getMotorVoltage, float());
   MOCK_METHOD0(getMotorCurrent, float());
 
-  MOCK_METHOD1(actuateRadians, void(double));
-  MOCK_METHOD1(actuateTorque, void(double));
+  MOCK_METHOD1(actuateRadians, void(float));
+  MOCK_METHOD1(actuateTorque, void(float));
 
   MOCK_METHOD2(initSdo, bool(march::SdoSlaveInterface& sdo, int cycle_time));
   MOCK_METHOD1(reset, void(march::SdoSlaveInterface&));
